requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
configmypy==0.2.0
cycler==0.12.1
einops==0.8.2
filelock==3.20.0
fonttools==4.62.1
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.46
h5py==3.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.2
matplotlib>=3.0
mpmath==1.3.0
networkx>=3.0
neuraloperator==2.0.0
numpy==2.3.5
nvidia-ml-py==13.590.48
opt_einsum==3.4.0
pillow==12.0.0
platformdirs==4.9.4
pluggy==1.6.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.19.1
scipy==1.17.1
sentry-sdk==2.55.0
six==1.17.0
smmap==5.0.3
sympy==1.14.0
tensorly==0.9.0
tensorly-torch==0.5.0
the_well==1.2.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.1
zencfg==0.7.0
